we can now apply transformations

+++ a1/Cube.py
@@ -68,17 +68,17 @@ class Cube:
         # create edges
         for first in range(0, n+1):
             for second in range(0, n+1):
-                i = (2.0 / float(n)) * first - 1.0
-                j = (2.0 / float(n)) * second - 1.0
+                i = float((2.0 / float(n)) * first - 1.0)
+                j = float((2.0 / float(n)) * second - 1.0)
 
-                i_plus1 = (2.0 / float(n)) * (1.0 + first) - 1.0
-                j_plus1 = (2.0 / float(n)) * (1.0 + second) - 1.0
+                i_plus1 = float((2.0 / float(n)) * (1.0 + first) - 1.0)
+                j_plus1 = float((2.0 / float(n)) * (1.0 + second) - 1.0)
 
                 # z = 1
-                coordinates = (i, j, 1)
-                horizontal  = (i_plus1, j, 1)
-                vertical    = (i, j_plus1, 1)
-                diagonal    = (i_plus1, j_plus1, 1)
+                coordinates = (i, j, 1.0)
+                horizontal  = (i_plus1, j, 1.0)
+                vertical    = (i, j_plus1, 1.0)
+                diagonal    = (i_plus1, j_plus1, 1.0)
                 if first != n and e_dict.get( (coordinates, horizontal) ) == None:
                     e_dict[ (coordinates, horizontal) ] = True
                     m.edges.append( (coordinates, horizontal) )
@@ -92,10 +92,10 @@ class Cube:
                     m.edges.append( (coordinates, diagonal) )
 
                 # z = -1
-                coordinates = (i, j, -1)
-                horizontal  = (i_plus1, j, -1)
-                vertical    = (i, j_plus1, -1)
-                diagonal    = (i_plus1, j_plus1, -1)
+                coordinates = (i, j, -1.0)
+                horizontal  = (i_plus1, j, -1.0)
+                vertical    = (i, j_plus1, -1.0)
+                diagonal    = (i_plus1, j_plus1, -1.0)
                 if first != n and e_dict.get( (coordinates, horizontal) ) == None:
                     e_dict[ (coordinates, horizontal) ] = True
                     m.edges.append( (coordinates, horizontal) )
@@ -109,10 +109,10 @@ class Cube:
                     m.edges.append( (coordinates, diagonal) )
 
                 # y = 1 
-                coordinates = (i, 1, j)
-                horizontal  = (i_plus1, 1, j)
-                vertical    = (i, 1, j_plus1)
-                diagonal    = (i_plus1, 1, j_plus1)
+                coordinates = (i, 1.0, j)
+                horizontal  = (i_plus1, 1.0, j)
+                vertical    = (i, 1.0, j_plus1)
+                diagonal    = (i_plus1, 1.0, j_plus1)
                 if first != n and e_dict.get( (coordinates, horizontal) ) == None:
                     e_dict[ (coordinates, horizontal) ] = True
                     m.edges.append( (coordinates, horizontal) )
@@ -126,10 +126,10 @@ class Cube:
                     m.edges.append( (coordinates, diagonal) )
 
                 # y = -1
-                coordinates = (i, -1, j)
-                horizontal  = (i_plus1, -1, j)
-                vertical    = (i, -1, j_plus1)
-                diagonal    = (i_plus1, -1, j_plus1)
+                coordinates = (i, -1.0, j)
+                horizontal  = (i_plus1, -1.0, j)
+                vertical    = (i, -1.0, j_plus1)
+                diagonal    = (i_plus1, -1.0, j_plus1)
                 if first != n and e_dict.get( (coordinates, horizontal) ) == None:
                     e_dict[ (coordinates, horizontal) ] = True
                     m.edges.append( (coordinates, horizontal) )
@@ -143,10 +143,10 @@ class Cube:
                     m.edges.append( (coordinates, diagonal) )
 
                 # x = 1
-                coordinates = (1, i, j)
-                horizontal  = (1, i_plus1, j)
-                vertical    = (1, i, j_plus1)
-                diagonal    = (1, i_plus1, j_plus1)
+                coordinates = (1.0, i, j)
+                horizontal  = (1.0, i_plus1, j)
+                vertical    = (1.0, i, j_plus1)
+                diagonal    = (1.0, i_plus1, j_plus1)
                 if first != n and e_dict.get( (coordinates, horizontal) ) == None:
                     e_dict[ (coordinates, horizontal) ] = True
                     m.edges.append( (coordinates, horizontal) )
@@ -160,10 +160,10 @@ class Cube:
                     m.edges.append( (coordinates, diagonal) )
             
                 # x = -1
-                coordinates = (-1, i, j)
-                horizontal  = (-1, i_plus1, j)
-                vertical    = (-1, i, j_plus1)
-                diagonal    = (-1, i_plus1, j_plus1)
+                coordinates = (-1.0, i, j)
+                horizontal  = (-1.0, i_plus1, j)
+                vertical    = (-1.0, i, j_plus1)
+                diagonal    = (-1.0, i_plus1, j_plus1)
                 if first != n and e_dict.get( (coordinates, horizontal) ) == None:
                     e_dict[ (coordinates, horizontal) ] = True
                     m.edges.append( (coordinates, horizontal) )
